JackBaboon

Julia executor with limited concurrency.

API

JackBaboon.Executor — Type

Executor(;
    pool         :: Symbol  = :default,
    capacity     :: Integer = 8,
    concurrently :: Integer = 1,
)

Create and run executor with limited concurrency.

Arguments

pool: thread pool; supported: :default, :interactive.

capacity: maximum number of queued jobs; new jobs are rejected when queue is full.

concurrently: maximum number of jobs executing concurrently.

JackBaboon.execute! — Function

execute!(@nospecialize(f), executor::Executor)

Synchronous concurrently execution.

Examples

julia> result = execute!(executor) do cancel_token
           do_work()
       end;

julia> result = execute!(executor) do cancel_token
           error("Job error")
       end;
ERROR: Job error
Stacktrace:
 [1] error()
 ...

JackBaboon.submit! — Function

submit!(@nospecialize(f), executor::Executor)::Handle

Asynchronous concurrently execution.

Examples

julia> handle = submit!(executor) do cancel_token
           do_work()
       end;
    
julia> result = fetch(handle);

julia> handle = submit!(executor) do cancel_token
           error("Job error")
       end;

julia> fetch(handle)
ERROR: Job error
Stacktrace:
 [1] error(s::String)
 ...

Base.isopen — Method

isopen(executor::Executor)::Bool

Returns true if executor is ready to accept jobs.

Base.close — Method

close(executor::Executor)::Executor

Graceful shutdown; immediately stop accepting new jobs; accepted jobs are finalize.

JackBaboon.isclosed — Function

isclosed(executor::Executor)::Bool

Returns true if executor is closed.

JackBaboon.iscrashed — Function

iscrashed(executor::Executor)::Bool

Returns true if executor is crashed.

JackBaboon.iscancelrequested — Function

iscancelrequested(token::CancelToken)::Bool

Return true if stop has been requested for the job.

Examples

julia> executor = Executor();

julia> handle = submit!(executor) do cancel_token
        while ! iscancelrequested(cancel_token)
            do_work()
            yield()
        end
        nothing
    end;

julia> stop!(handle);

julia> wait(handle)

julia> iscanceled(handle)
true

JackBaboon.isqueued — Function

isqueued(handle::Handle)::Bool

Returns true if job in executor queue.

JackBaboon.ispending — Function

ispending(handle::Handle)::Bool

Returns true if job is pending execution.

JackBaboon.iscanceled — Function

iscanceled(handle::Handle)::Bool

Returns true if job was canceled and task is completed.

JackBaboon.isrunning — Function

isrunning(handle::Handle)::Bool

Returns true if job task is executing.

JackBaboon.iscompleted — Function

iscompleted(handle::Handle)::Bool

Returns true if job task completed successfully.

JackBaboon.stop! — Function

stop!(handle::Handle)::Handle

Job stopping; job completion is a user responsibility.

JackBaboon.isstopping — Function

isstopping(handle::Handle)::Bool

Returns true if running job stopping in progress.

JackBaboon.isstopped — Function

isstopped(handle::Handle)::Bool

Returns true if job is stopped.

JackBaboon.isfailed — Function

isfailed(handle::Handle)::Bool

Returns true if job task is failed.

JackBaboon.isfinal — Function

isfinal(handle::Handle)::Bool

Returns true if job in final state.

Base.wait — Method

wait(handle::Handle; throw=true)::Nothing

Wait job finalize.

Base.fetch — Method

fetch(handle::Handle)::Any

Wait job finalize and fetch result or throw error.

JackBaboon.ExecutorClosedError — Type

ExecutorClosedError(msg::AbstractString)

Thrown on a job submitting to a closed executor.

JackBaboon.ExecutorRejectedError — Type

ExecutorRejectedError(msg::AbstractString)

Thrown on a job submitting to an executor with a full queue.

JackBaboon.JobCancelledError — Type

JobCancelledError(msg::AbstractString)

Thrown on fetch result from cancelling or cancelled job.

JackBaboon.ExecutorInternalError — Type

ExecutorInternalError(msg::AbstractString)

Thrown on an internal executor error.